refactor(shared): move the fs-backed document loader factory out of the loader interface module
