+++ cli/src/cli.spec.ts
@@ -14,7 +14,7 @@ let templateModule: typeof import('./command-helpers/template');
 let optionsModule: typeof import('./command-helpers/generate-options');
 let diffModule: typeof import('./command-helpers/diff');
 let hubCommandsModule: typeof import('./command-helpers/hub-commands');
-let documentLoaderModule: typeof import('../../shared/src/document-loader/document-loader');
+let documentLoaderModule: typeof import('../../shared/src/document-loader/node-document-loader');
 let setupCLI: typeof import('./cli').setupCLI;
 let cliConfigModule: typeof import('./cli-config');
 
@@ -31,7 +31,7 @@ describe('CLI Commands', () => {
         templateModule = await import('./command-helpers/template');
         optionsModule = await import('./command-helpers/generate-options');
         diffModule = await import('./command-helpers/diff');
-        documentLoaderModule = await import('../../shared/src/document-loader/document-loader');
+        documentLoaderModule = await import('../../shared/src/document-loader/node-document-loader');
 
         vi.spyOn(calmShared, 'runGenerate').mockResolvedValue(undefined);
         vi.spyOn(calmShared.TemplateProcessor.prototype, 'processTemplate').mockResolvedValue(undefined);
