+++ shared/src/document-loader/node-document-loader.spec.ts
@@ -1,0 +1,201 @@
+import { AuthPlugin } from '../auth/auth-plugin';
+import { CALM_META_SCHEMA_DIRECTORY } from '../consts';
+import { CALM_DOCUMENT_TYPES_LIST, isValidCalmDocumentType } from '@finos/calm-models/types';
+import { DocumentLoaderOptions } from './document-loader';
+import { buildDocumentLoader } from './node-document-loader';
+
+const mocks = vi.hoisted(() => {
+    return {
+        fsDocLoader: vi.fn(function () { return {
+            initialise: vi.fn(),
+            loadMissingDocument: vi.fn()
+        }; }),
+        calmHubDocLoader: vi.fn(function () { return {
+            initialise: vi.fn(),
+            loadMissingDocument: vi.fn()
+        }; }),
+        mappedDocLoader: vi.fn(function () { return {
+            initialise: vi.fn(),
+            loadMissingDocument: vi.fn()
+        }; }),
+        directDocLoader: vi.fn(function () { return {
+            initialise: vi.fn(),
+            loadMissingDocument: vi.fn()
+        }; }),
+        workspaceDocLoader: vi.fn(function () { return {
+            initialise: vi.fn(),
+            loadMissingDocument: vi.fn()
+        }; })
+    };
+});
+
+
+vi.mock('./file-system-document-loader', () => {
+    return {
+        FileSystemDocumentLoader: mocks.fsDocLoader
+    };
+});
+
+vi.mock('./calmhub-document-loader', () => {
+    return {
+        CalmHubDocumentLoader: mocks.calmHubDocLoader
+    };
+});
+
+vi.mock('./mapped-document-loader', () => {
+    return {
+        MappedDocumentLoader: mocks.mappedDocLoader
+    };
+});
+
+vi.mock('./direct-url-document-loader', () => {
+    return {
+        DirectUrlDocumentLoader: mocks.directDocLoader
+    };
+});
+
+vi.mock('./workspace-document-loader', () => {
+    return {
+        WorkspaceDocumentLoader: mocks.workspaceDocLoader
+    };
+});
+
+describe('DocumentLoader', () => {
+    beforeEach(() => {
+        vi.clearAllMocks();
+        vi.resetModules();
+    });
+
+    it('should create a FileSystemDocumentLoader', () => {
+
+        const docLoaderOpts: DocumentLoaderOptions = {
+            schemaDirectoryPath: 'schemas'
+        };
+
+        buildDocumentLoader(docLoaderOpts);
+
+        expect(mocks.fsDocLoader).toHaveBeenCalledWith([CALM_META_SCHEMA_DIRECTORY, 'schemas'], false, process.cwd());
+    });
+
+    it('should not create a WorkspaceDocumentLoader when workspaceBundlePath is absent', () => {
+        buildDocumentLoader({ schemaDirectoryPath: 'schemas' });
+        expect(mocks.workspaceDocLoader).not.toHaveBeenCalled();
+    });
+
+    it('should create a WorkspaceDocumentLoader when workspaceBundlePath is provided', () => {
+        const docLoaderOpts: DocumentLoaderOptions = {
+            workspaceBundlePath: '/repo/.calm-workspace/bundles/default'
+        };
+
+        buildDocumentLoader(docLoaderOpts);
+
+        expect(mocks.workspaceDocLoader).toHaveBeenCalledWith('/repo/.calm-workspace/bundles/default', false);
+    });
+
+    it('should create a CalmHubDocumentLoader when calmHubUrl is defined in loader options', () => {
+
+        const docLoaderOpts: DocumentLoaderOptions = {
+            calmHubUrl: 'https://example.com'
+        };
+
+        buildDocumentLoader(docLoaderOpts);
+
+        expect(mocks.calmHubDocLoader).toHaveBeenCalledWith('https://example.com', false, undefined);
+    });
+    
+    it('should pass authplugin to CalmHubDocumentLoader', () => {
+
+        const mockAuthPlugin: AuthPlugin = {
+            getAuthHeaders: vi.fn()
+        };
+
+        const docLoaderOpts: DocumentLoaderOptions = {
+            calmHubUrl: 'https://example.com',
+            authPlugin: mockAuthPlugin
+        };
+
+        buildDocumentLoader(docLoaderOpts);
+
+        expect(mocks.calmHubDocLoader).toHaveBeenCalledWith('https://example.com', false, mockAuthPlugin);
+    });
+
+    it('should pass allowedRemoteHosts to DirectUrlDocumentLoader when provided', () => {
+        const docLoaderOpts: DocumentLoaderOptions = {
+            allowedRemoteHosts: ['schemas.example.com']
+        };
+
+        buildDocumentLoader(docLoaderOpts);
+
+        expect(mocks.directDocLoader).toHaveBeenCalledWith(false, undefined, ['schemas.example.com']);
+    });
+
+    it('should create a MappedDocumentLoader when urlToLocalMap is provided', () => {
+        const urlMap = new Map([
+            ['https://example.com/schema.json', 'local/schema.json']
+        ]);
+
+        const docLoaderOpts: DocumentLoaderOptions = {
+            urlToLocalMap: urlMap
+        };
+
+        buildDocumentLoader(docLoaderOpts);
+
+        expect(mocks.mappedDocLoader).toHaveBeenCalledWith(urlMap, process.cwd(), false);
+    });
+
+    it('should create a MappedDocumentLoader when basePath is provided', () => {
+        const docLoaderOpts: DocumentLoaderOptions = {
+            basePath: '/project/patterns'
+        };
+
+        buildDocumentLoader(docLoaderOpts);
+
+        expect(mocks.mappedDocLoader).toHaveBeenCalledWith(new Map(), '/project/patterns', false);
+    });
+
+    it('should create a MappedDocumentLoader with both urlToLocalMap and basePath', () => {
+        const urlMap = new Map([
+            ['https://example.com/schema.json', 'local/schema.json']
+        ]);
+
+        const docLoaderOpts: DocumentLoaderOptions = {
+            urlToLocalMap: urlMap,
+            basePath: '/custom/base'
+        };
+
+        buildDocumentLoader(docLoaderOpts);
+
+        expect(mocks.mappedDocLoader).toHaveBeenCalledWith(urlMap, '/custom/base', false);
+    });
+
+    it('should not create a MappedDocumentLoader when neither urlToLocalMap nor basePath provided', () => {
+        const docLoaderOpts: DocumentLoaderOptions = {};
+
+        buildDocumentLoader(docLoaderOpts);
+
+        expect(mocks.mappedDocLoader).not.toHaveBeenCalled();
+    });
+
+    it('should not create a MappedDocumentLoader when urlToLocalMap is empty and no basePath', () => {
+        const docLoaderOpts: DocumentLoaderOptions = {
+            urlToLocalMap: new Map()
+        };
+
+        buildDocumentLoader(docLoaderOpts);
+
+        expect(mocks.mappedDocLoader).not.toHaveBeenCalled();
+    });
+});
+
+describe('isValidCalmDocumentType', () => {
+    it.each(CALM_DOCUMENT_TYPES_LIST)('returns true for the valid document type %s', (type) => {
+        expect(isValidCalmDocumentType(type)).toBe(true);
+    });
+
+    it.each(['unknown', 'architectures', 'Pattern', '', 'foo'])(
+        'returns false for the invalid document type %s',
+        (type) => {
+            expect(isValidCalmDocumentType(type)).toBe(false);
+        }
+    );
+});
