+++ shared/src/index.ts
@@ -58,7 +58,8 @@ export { C4Model } from './docify/graphing/c4.js';
 export { CalmRelationshipGraph } from './docify/graphing/relationship-graph.js';
 export { ValidationOutcome } from './commands/validate/validation.output.js';
 export { setWidgetLogger, type WidgetLogger } from '@finos/calm-widgets';
-export { buildDocumentLoader, DocumentLoader, DocumentLoaderOptions } from './document-loader/document-loader.js';
+export { DocumentLoader, DocumentLoaderOptions, DocumentLoadError, assertJsonObject, CALM_HUB_PROTOS } from './document-loader/document-loader.js';
+export { buildDocumentLoader } from './document-loader/node-document-loader.js';
 export { FileSystemDocumentLoader } from './document-loader/file-system-document-loader.js';
 export { WorkspaceDocumentLoader } from './document-loader/workspace-document-loader.js';
 export * from './document-loader/loading-helpers.js';
